계정 신청 구현(암호화 주민번호 null일때 미구현)

+++ src/main/java/com/example/approval24/controller/AuthorityController.java
@@ -106,7 +106,7 @@ public class AuthorityController {
         model.addAttribute("authorityDepartments", authorityDepartments);
 
         // 미할당 부서 목록 조회
-        Set<Integer> assignedDeptIds = authorityDepartments.stream()
+        Set<Long> assignedDeptIds = authorityDepartments.stream()
                 .map(DeptDTO::getDeptId)
                 .collect(Collectors.toSet());
